+++ src/Model/Table/PostsTable.php
@@ -24,7 +24,7 @@ use Cake\ORM\Query;
 use Cake\ORM\RulesChecker;
 use MeCms\Model\Entity\Post;
 use MeCms\Model\Table\AppTable;
-use MeCms\Model\Table\Traits\GetPreviewFromTextTrait;
+use MeCms\Model\Table\Traits\GetPreviewsFromTextTrait;
 use MeCms\Model\Table\Traits\IsOwnedByTrait;
 use MeCms\Model\Table\Traits\NextToBePublishedTrait;
 
@@ -45,7 +45,7 @@ use MeCms\Model\Table\Traits\NextToBePublishedTrait;
  */
 class PostsTable extends AppTable
 {
-    use GetPreviewFromTextTrait;
+    use GetPreviewsFromTextTrait;
     use IsOwnedByTrait;
     use LocatorAwareTrait;
     use NextToBePublishedTrait;
@@ -143,7 +143,7 @@ class PostsTable extends AppTable
      * @return void
      * @since 2.17.0
      * @uses MeCms\Model\Table\AppTable::beforeSave()
-     * @uses MeCms\Model\Table\Traits\GetPreviewFromTextTrait::getPreview()
+     * @uses MeCms\Model\Table\Traits\GetPreviewsFromTextTrait::getPreview()
      */
     public function beforeSave(Event $event, EntityInterface $entity, ArrayObject $options)
     {